requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
reportlab==4.2.2
Pillow==10.4.0
python-dotenv==1.0.1
# rdkit-pypi==2022.9.5
